feat - 完善点赞与评论功能，完善token逻辑，添加房源预约逻辑及功能；所有接口请求方式改为post

+++ starfish_server/src/main/java/com/zoe/starfish_server/service/DianZanService.java
@@ -26,7 +26,11 @@ public class DianZanService {
      * @return
      */
     public Boolean cancel(Integer type, Long newsId, Long pingLunId) {
-        Long id = getDianZan(type, newsId, pingLunId).getId();
+        DianZan dianZan = getDianZan(type, newsId, pingLunId);
+        if (dianZan == null) {
+            return false;
+        }
+        Long id = dianZan.getId();
         //取消点赞
         return mapper.deleteByPrimaryKey(id) == 1;
     }
